requirements.txt
alembic==1.2.1
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-Classful==0.14.2
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
